junitmock/src/test/java/com/demo/junitmock/servlet/GetIndexPageServletTest.java
package com.demo.junitmock.servlet;

import org.junit.jupiter.api.Test;

import javax.servlet.RequestDispatcher;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import java.io.IOException;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

class GetIndexPageServletTest {

    private final static String path = "/static/index.html";

    @Test
    void callDoGetServletReturnIndexPage() throws ServletException, IOException {
        final GetIndexPageServlet servlet = new GetIndexPageServlet();

        final HttpServletRequest request = mock(HttpServletRequest.class);
        final HttpServletResponse response = mock(HttpServletResponse.class);
        final RequestDispatcher dispatcher = mock(RequestDispatcher.class);
        when(request.getRequestDispatcher(path)).thenReturn(dispatcher);
        servlet.doGet(request, response);
        verify(request, times(1)).getRequestDispatcher(path);
        verify(request, never()).getSession();
        verify(dispatcher).forward(request,response);


    }
}